# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Roleva API",
    "description": "Evidence-based resume analysis.",
    "version": "0.1.0"
  },
  "paths": {
    "/analyze": {
      "post": {
        "summary": "Analyze",
        "description": "Analyse one resume against one job description.\n\nSynchronous. At fewer than a hundred analyses a day there is no queue worth\nrunning, and a request that returns the finished report is far simpler to\nreason about than a job id the client has to poll.",
        "operationId": "analyze_analyze_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/analyze/stream": {
      "post": {
        "summary": "Analyze Stream",
        "description": "The same analysis, with progress events while it runs.\n\nEverything that can be rejected is rejected *before* the stream opens, so a\nquota error is still an HTTP 429 the client can handle normally. Once the\nfirst byte is sent the status is 200 for good, and any later failure has to\ntravel as an error frame instead.",
        "operationId": "analyze_stream_analyze_stream_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_stream_analyze_stream_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/analyses": {
      "get": {
        "summary": "List Analyses",
        "description": "The signed-in user's analysis history, newest first.",
        "operationId": "list_analyses_analyses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Analyses Analyses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analyses/{analysis_id}": {
      "get": {
        "summary": "Get Analysis",
        "operationId": "get_analysis_analyses__analysis_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Analysis",
        "description": "Delete an analysis permanently.\n\nNo soft delete. When somebody asks for their resume data to be removed, a\nrow marked `deleted = true` is not removal.",
        "operationId": "delete_analysis_analyses__analysis_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/me/quota": {
      "get": {
        "summary": "Quota Remaining",
        "description": "How many analyses the user has left today.\n\nRead without incrementing \u2014 `bump_rate_limit` is the only thing that counts,\nand asking how many you have left must not spend one. The counter is written\nby that function and simply read here.",
        "operationId": "quota_remaining_me_quota_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Quota Remaining Me Quota Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/me/export": {
      "get": {
        "summary": "Export My Data",
        "description": "Everything Roleva holds about this user, as JSON.\n\nA data-export endpoint is only honest if it returns *everything*, so this\nreads every user-owned table rather than a curated subset. It runs with the\ncaller's own token, so RLS decides what comes back \u2014 which means the export\ncannot accidentally include somebody else's row even if a filter were wrong.\n\n`score_samples` is deliberately absent, and its absence is the point: those\nrows carry no user id, so there is nothing to attribute to anyone. Saying so\nhere is more useful than silently omitting them.",
        "operationId": "export_my_data_me_export_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Export My Data Me Export Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/me": {
      "get": {
        "summary": "Me",
        "description": "Confirms a token is valid. Returns the user id only \u2014 the backend has no\nreason to echo an email address back to a client that already has it.",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Me Me Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "summary": "Delete My Account",
        "description": "Delete the account and everything attached to it.\n\nUses the service-role key, which is the one place a user-owned deletion\nlegitimately needs it: removing a row from `auth.users` is an admin\noperation, and every table's foreign key cascades from there. So one call\nremoves the profile, the resumes, the job targets, the analyses and the\nshare links.\n\nA real delete, not a flag. Someone asking for their resume data to be\nremoved is not asking for a column to be set to true.",
        "operationId": "delete_my_account_me_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisReport": {
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "resume": {
            "$ref": "#/components/schemas/ResumeDocument"
          },
          "job": {
            "$ref": "#/components/schemas/JobTarget"
          },
          "matches": {
            "$ref": "#/components/schemas/MatchReport"
          },
          "ats": {
            "$ref": "#/components/schemas/AtsReport"
          },
          "scores": {
            "$ref": "#/components/schemas/ScoreReport"
          },
          "headline": {
            "default": "",
            "title": "Headline",
            "type": "string"
          },
          "verdict": {
            "default": "",
            "title": "Verdict",
            "type": "string"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "title": "Strengths",
            "type": "array"
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "title": "Weaknesses",
            "type": "array"
          },
          "section_feedback": {
            "items": {
              "$ref": "#/components/schemas/SectionFeedback"
            },
            "title": "Section Feedback",
            "type": "array"
          },
          "bullet_suggestions": {
            "items": {
              "$ref": "#/components/schemas/BulletSuggestion"
            },
            "title": "Bullet Suggestions",
            "type": "array"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "rubric_version": {
            "title": "Rubric Version",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "degraded_stages": {
            "items": {
              "$ref": "#/components/schemas/Stage"
            },
            "title": "Degraded Stages",
            "type": "array"
          }
        },
        "required": [
          "id",
          "status",
          "created_at",
          "resume",
          "job",
          "matches",
          "ats",
          "scores",
          "rubric_version",
          "prompt_version"
        ],
        "title": "AnalysisReport",
        "type": "object"
      },
      "AnalysisStatus": {
        "enum": [
          "pending",
          "running",
          "complete",
          "partial",
          "failed"
        ],
        "title": "AnalysisStatus",
        "type": "string"
      },
...
      "Body_analyze_analyze_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "The resume, as a PDF"
          },
          "job_description": {
            "type": "string",
            "title": "Job Description",
            "description": "The full job posting text"
          }
        },
        "type": "object",
        "required": [
          "file",
          "job_description"
        ],
        "title": "Body_analyze_analyze_post"
      },
      "Body_analyze_stream_analyze_stream_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "The resume, as a PDF"
          },
          "job_description": {
            "type": "string",
            "title": "Job Description",
            "description": "The full job posting text"
          }
        },
        "type": "object",
        "required": [
          "file",
          "job_description"
        ],
        "title": "Body_analyze_stream_analyze_stream_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResumeDocument": {
        "additionalProperties": false,
        "description": "The structured resume. Analyzer reads it; the future builder will edit it.",
        "properties": {
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "contact": {
            "$ref": "#/components/schemas/ContactInfo"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Bullet"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "experience": {
            "items": {
              "$ref": "#/components/schemas/ExperienceItem"
            },
            "title": "Experience",
            "type": "array"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationItem"
            },
            "title": "Education",
            "type": "array"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectItem"
            },
            "title": "Projects",
            "type": "array"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillMention"
            },
            "title": "Skills",
            "type": "array"
          },
          "certifications": {
            "items": {
              "$ref": "#/components/schemas/CredentialItem"
            },
            "title": "Certifications",
            "type": "array"
          },
          "awards": {
            "items": {
              "$ref": "#/components/schemas/CredentialItem"
            },
            "title": "Awards",
            "type": "array"
          },
          "publications": {
            "items": {
              "$ref": "#/components/schemas/CredentialItem"
            },
            "title": "Publications",
            "type": "array"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/DetectedSection"
            },
            "title": "Sections",
            "type": "array"
          },
          "page_count": {
            "default": 0,
            "minimum": 0,
            "title": "Page Count",
            "type": "integer"
          },
          "parse_confidence": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Parse Confidence",
            "type": "number"
          },
          "parse_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Parse Warnings",
            "type": "array"
          }
        },
        "title": "ResumeDocument",
        "type": "object"
      },
...
      "Stage": {
        "description": "Pipeline stages. Names are user-visible in the progress UI, so they\ndescribe work the user cares about, not internal function names.",
        "enum": [
          "validating",
          "extracting",
          "structuring",
          "reading_job",
          "matching",
          "checking_ats",
          "assessing_quality",
          "scoring",
          "writing_advice",
          "done"
        ],
        "title": "Stage",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}